LayoutTests/fast/text/pre-wrap-trailing-tab-expected.txt

This test ensures CyberKit renders the trailing whitespace properly. You should see PASS thrice below.

test 0: PASS
test 1: PASS
test 2: PASS
